Install playwright to scrollspy
