+++ packages/scroll-spy/__tests__/playwright/playwright.spec.js
@@ -1,0 +1,23 @@
+const { test, expect } = require('@playwright/test');
+import AxeBuilder from '@axe-core/playwright';
+
+let tabKey;
+
+test.beforeEach(async ({ page }, testInfo) => {
+	await page.goto('/');
+	tabKey = testInfo.project.use.defaultBrowserType === 'webkit'
+			? "Alt+Tab"
+			: "Tab";
+
+	await page.keyboard.press(tabKey);
+	await page.keyboard.press('Enter');
+});
+
+
+test.describe('Skip > Axe', { tag: '@reduced'}, () => {
+	test('Should not have any automatically detectable accessibility issues', async ({ page }) => {	
+		const accessibilityScanResults = await new AxeBuilder({ page }).analyze(); 
+		expect(accessibilityScanResults.violations).toEqual([]);
+	});
+});
+
