+++ packages/scroll-spy/example/src/index.html
@@ -1,5 +1,5 @@
 <!doctype html>
-<html>
+<html lang="en">
 	<head>
     <title>StormID</title>
     <style>
@@ -80,7 +80,7 @@
             <a class="js-scroll-spy" href="#section3">Section 3</a>
         </nav>
     </header>
-    <div class="container">
+    <main class="container">
         <h1>Scroll Spy</h1>
         <h2>Example</h2>
         <p>Scroll down to see the menu to highlight the current section and to see example code.</p>
@@ -93,5 +93,6 @@
         <section id="section3" style="height:500px">
             Section 3
         </section>
+    </main>
 </body>
 </html>
