+++ packages/scroll-spy/package.json
@@ -27,9 +27,9 @@
   ],
   "scripts": {
     "build": "microbundle --name ScrollSpy",
-    "test": "jest --coverage",
-    "dev": "webpack-dev-server --config example/webpack.config.js",
-    "prod": "webpack --config example/webpack.config.js --mode production",
+    "test": "jest --coverage & npx playwright test",
+    "dev": "webpack-dev-server --config tools/webpack.config.js",
+    "prod": "webpack --config tools/webpack.config.js --mode production",
     "prepublish": "npm run -s build"
   },
   "dependencies": {
